move read-protocol-state-blocks command to common
