+++ admin/commands/common/read_protocol_state_blocks.go
@@ -1,4 +1,4 @@
-package consensus
+package common
 
 import (
 	"context"
